split out helpers
